requirements.txt
vk-api>=4.5.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
